requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
python-docx==0.8.11
python-dotenv==1.0.0